+++ test/builders/json_animated_align_builder_test.dart
@@ -1,0 +1,20 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:json_dynamic_widget/json_dynamic_widget.dart';
+
+void main() {
+  test('type', () {
+    const type = JsonAnimatedAlignBuilder.type;
+
+    expect(type, 'animated_align');
+    expect(JsonWidgetRegistry.instance.getWidgetBuilder(type) != null, true);
+    expect(
+      JsonWidgetRegistry.instance.getWidgetBuilder(type)(
+        {
+          'alignment': 'bottomLeft',
+          'duration': 1000,
+        },
+      ) is JsonAnimatedAlignBuilder,
+      true,
+    );
+  });
+}