+++ test/builders/json_animated_padding_builder_test.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/painting.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:json_dynamic_widget/json_dynamic_widget.dart';
+
+void main() {
+  test('type', () {
+    const type = JsonAnimatedPaddingBuilder.type;
+
+    expect(type, 'animated_padding');
+    expect(JsonWidgetRegistry.instance.getWidgetBuilder(type) != null, true);
+    expect(
+      JsonWidgetRegistry.instance.getWidgetBuilder(type)(
+        {
+          'duration': 1000,
+          'padding': EdgeInsets.zero,
+        },
+      ) is JsonAnimatedPaddingBuilder,
+      true,
+    );
+  });
+}